+++ src/python/test/test_dxclient.py
@@ -6268,6 +6268,33 @@ class TestDXClientNewUser(DXTestCase):
             with self.assertRaisesRegex(subprocess.CalledProcessError,
                                          "DXCLIError"):
                 run(" ".join([cmd, invalid_opts]))
+    
+    def test_create_user_on_behalf_of(self):
+        username, email = generate_unique_username_email()
+        first = "Asset"
+        cmd = "dx new user"
+        baseargs = "--username {u} --email {e} --first {f}".format(u=username, e=email, f=first)
+        user_id = run(" ".join([cmd, baseargs,"--on-behalf-of {o} --brief".format(o=self.org_id)])).strip()
+        self._assert_user_desc(user_id, {"first": first})
+    
+    def test_create_user_on_behalf_of_negative(self):
+        username, email = generate_unique_username_email()
+        first = "Asset2"
+        cmd = "dx new user"
+        baseargs = "--username {u} --email {e} --first {f}".format(u=username, e=email, f=first)
+    
+        # no org specified
+        with self.assertRaisesRegex(subprocess.CalledProcessError,
+                                    "error: argument --on-behalf-of: expected one argument"):   
+            run(" ".join([cmd, baseargs,"--on-behalf-of" ]))
+        # creating user on behalf of org that does not exist 
+        with self.assertRaisesRegex(subprocess.CalledProcessError,
+                                        "ResourceNotFound"):
+            run(" ".join([cmd, baseargs,"--on-behalf-of org-does_not_exist"]))
+        # creating user for org in which the adder does not have ADMIN permissions
+        with self.assertRaisesRegex(subprocess.CalledProcessError,
+                                    "(PermissionDenied)|(ResourceNotFound)"):
+            run(" ".join([cmd, baseargs,"--on-behalf-of org-dnanexus"]))
 
 
     def test_self_signup_negative(self):