+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 
 ## Unreleased
 
+### Added
+
+* `--on-behalf-of <org>` argument for `dx new user`
+
 ## [327.1]
 
 ## Fixed